introduce route classes to structure service
