+++ server/src/main/scala/org/kisobran/top/TopListService.scala
@@ -1,28 +1,31 @@
 package org.kisobran.top
 
-import java.util.UUID
-
 import akka.http.scaladsl.server.directives.Credentials
 import akka.http.scaladsl.server.{Directives, Route}
-import org.kisobran.top.db.{Stats, TopListEntries, YtUtil}
-import org.kisobran.top.model.{Entry, Highlight}
-import org.kisobran.top.repository.{StatsRepository, TopListRepository}
-import org.kisobran.top.shared.SharedMessages
-import org.kisobran.top.twirl.Implicits._
-import Configuration._
 import com.github.blemale.scaffeine
 import com.github.blemale.scaffeine.Scaffeine
+import org.kisobran.top.Configuration._
+import org.kisobran.top.db.{TopListEntries, YtUtil}
+import org.kisobran.top.model.Highlight
 import org.kisobran.top.model.Highlight._
+import org.kisobran.top.repository.{StatsRepository, TopListRepository}
+import org.kisobran.top.routes.{AdminRoutes, ArchiveRoutes, VoteRoutes}
+import org.kisobran.top.shared.SharedMessages
+import org.kisobran.top.twirl.Implicits._
 import org.kisobran.top.util.LoggingSupport
 
+import scala.concurrent.ExecutionContext
 import scala.concurrent.duration._
-import scala.concurrent.{ExecutionContext, Future}
+
+object TopListService {
+  type EntryCache = scaffeine.AsyncLoadingCache[(Int, Int, Boolean, Int), Seq[TopListEntries]]
+}
 
 class TopListService(topListRepository: TopListRepository, statsRepository: StatsRepository)
                     (implicit executionContext: ExecutionContext)
   extends Directives with YtUtil with LoggingSupport {
 
-  val selectCache: scaffeine.AsyncLoadingCache[(Int, Int, Boolean, Int), Seq[TopListEntries]] =
+  val selectCache: TopListService.EntryCache =
     Scaffeine()
       .recordStats()
       .expireAfterWrite(40.minutes)
@@ -31,14 +34,14 @@ class TopListService(topListRepository: TopListRepository, statsRepository: Stat
       case (limit, offset, isEnabled, year) => topListRepository.select(limit, offset, isEnabled, year)
     }
 
-  val limit = 20
-  val currentYear = 2018
+  lazy val myUserPassAuthenticator: Credentials => Option[String] = {
+    case p@Credentials.Provided(id) if p.verify(adminPassword) => Some(id)
+    case _ => None
+  }
 
-  def myUserPassAuthenticator(credentials: Credentials): Option[String] =
-    credentials match {
-      case p@Credentials.Provided(id) if p.verify(adminPassword) => Some(id)
-      case _ => None
-    }
+  lazy val adminRoutes = new AdminRoutes(myUserPassAuthenticator, topListRepository)
+  lazy val archiveRoutes = new ArchiveRoutes(topListRepository, selectCache)
+  lazy val voteRoutes = new VoteRoutes(topListRepository, statsRepository)
 
   val route: Route = {
     pathSingleSlash {
@@ -66,31 +69,9 @@ class TopListService(topListRepository: TopListRepository, statsRepository: Stat
           }
         }
       } ~
-      pathPrefix("admin") {
-        authenticateBasic(realm = "secure site", myUserPassAuthenticator) { _ =>
-          parameters('id ?, 'year ?) { (maybeId, maybeYear) =>
-            get {
-              complete {
-                maybeId match {
-                  case Some(id) =>
-                    topListRepository.findTopList(id).map { lista =>
-                      org.kisobran.top.html.lista.render(lista, false, true)
-                    }
-                  case None =>
-                    topListRepository.select(
-                      limit = 100,
-                      offset = 0,
-                      isEnabled = false,
-                      maybeYear.map(_.toInt).getOrElse(currentYear)
-                    ).map { all =>
-                      org.kisobran.top.html.admin.render(all)
-                    }
-                }
-              }
-            }
-          }
-        }
-      } ~
+      archiveRoutes.route ~
+      adminRoutes.route ~
+      voteRoutes.route ~
       pathPrefix("update") {
         post {
           formFieldMap { formContent =>
@@ -129,22 +110,6 @@ class TopListService(topListRepository: TopListRepository, statsRepository: Stat
           org.kisobran.top.html.reco.render(Highlight.items)
        }
       } ~
-      pathPrefix("arhiva") {
-        parameters('year) { _year =>
-          parameters('page ?) { pageS =>
-            complete {
-              val page = pageS.map(_.toInt).getOrElse(0)
-              topListRepository.count(_year.toInt).flatMap { count =>
-                selectCache.get(limit, limit * page, true, _year.toInt).map { all =>
-                  val backPage = if (page >= 1) Some(page - 1) else None
-                  val forwardPage = if (count <= (page + 1) * limit) None else Some(page + 1)
-                  org.kisobran.top.html.arhiva.render(all, forwardPage, backPage, _year.toInt)
-                }
-              }
-            }
-          }
-        }
-      } ~
       pathPrefix("liste") {
         parameters('page) { pageS =>
           complete {
@@ -165,38 +130,7 @@ class TopListService(topListRepository: TopListRepository, statsRepository: Stat
         encodeResponse {
           getFromResource("public/" + file)
         }
-      } ~ {
-      pathPrefix("vote") {
-        post {
-          formFieldMap { formContent =>
-            complete {
-              val title = formContent.getOrElse("listName", s"untilted-${UUID.randomUUID()}")
-              val entries = (1 to 10).map { index =>
-                Entry(formContent(s"inputArtist$index"), formContent(s"inputSong$index"), index, index)
-              }
-
-              log.info(title)
-              log.info(entries.toString)
-
-              topListRepository.createTopList(
-                formContent.get("userEmail"),
-                entries,
-                title,
-                false,
-                2018
-              ).map { topListEntry =>
-                val insertOperation: Future[Seq[Stats]] = topListEntry.map { topList =>
-                  statsRepository.createStats(topList.id, entries)
-                }.getOrElse(Future.successful(Seq.empty))
-
-                insertOperation.map { _ =>
-                  org.kisobran.top.html.lista.render(topListEntry, true, false)
-                }
-              }
-            }
-          }
-        }
       }
-    }
+
   }
 }
